% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunction.R
\name{basic_mean}
\alias{basic_mean}
\title{Basic mean calculationdevtool}
\usage{
basic_mean(vec)
}
\arguments{
\item{vec}{A numeric vector}
}
\value{
A double
}
\description{
Basic mean calculationdevtool
}
\examples{
basic_mean(1:3)
}
